02-大数据离线计算/spkweek2/back/requirements.txt
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1613247308275/work
chardet @ file:///C:/ci/chardet_1607706910910/work
cryptography @ file:///C:/ci/cryptography_1616769432139/work
Django==2.2.5
django-cors-headers==3.7.0
future==0.18.2
greenlet @ file:///C:/ci/greenlet_1620913803175/work
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1617877486026/work
JayDeBeApi @ file:///tmp/build/80754af9/jaydebeapi_1601401763009/work
JPype1 @ file:///C:/ci/jpype1_1609767174277/work
mkl-fft==1.3.0
mkl-random @ file:///C:/ci/mkl_random_1618854593605/work
mkl-service==2.3.0
numpy @ file:///C:/ci/numpy_and_numpy_base_1620831380353/work
pandas @ file:///C:/ci/pandas_1618365631664/work
pure-sasl @ file:///tmp/build/80754af9/pure-sasl_1600458057090/work
py4j==0.10.7
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyHive==0.6.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
PySocks @ file:///C:/ci/pysocks_1594394709107/work
pyspark==2.4.0
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==5.4.1
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
sasl==0.2.1
six @ file:///C:/ci/six_1605205426665/work
SQLAlchemy @ file:///C:/ci/sqlalchemy_1620712587153/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
thrift==0.13.0
thrift-sasl==0.4.2
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
